+++ e3nn/math/__init__.py
@@ -1,4 +1,4 @@
-from .linalg import *  # noqa
-from .normalize_activation import *  # noqa
-from .soft_one_hot import *  # noqa
-from .soft_heaviside import *  # noqa
+from ._linalg import *  # noqa
+from ._normalize_activation import *  # noqa
+from ._soft_unit_step import *  # noqa
+from ._soft_one_hot_linspace import *  # noqa