+++ e3nn/math/_soft_one_hot_linspace.py
@@ -1,24 +1,24 @@
 import math
 import torch
 
-from .soft_heaviside import soft_unit_step
+from e3nn.math import soft_unit_step
 
 
-def soft_one_hot_linspace(x, start, end, number, basis='gaussian', endpoint=True):
+def soft_one_hot_linspace(x, start, end, number, basis=None, cutoff=None):
     r"""Projection on a basis of functions
 
-    Returns a set of :math:`\{y_i\}_{i=1}^N`,
+    Returns a set of :math:`\{y_i(x)\}_{i=1}^N`,
 
     .. math::
 
-        y_i = \frac{1}{Z} f_i(x)
+        y_i(x) = \frac{1}{Z} f_i(x)
 
     where :math:`x` is the input and :math:`f_i` is the ith basis function.
-    :math:`Z` is set such that,
+    :math:`Z` is a constant defined (if possible) such that,
 
     .. math::
 
-        \sum_{i=1}^N y_i^2 \approx 1
+        \langle \sum_{i=1}^N y_i(x)^2 \rangle_x \approx 1
 
     See the last plot below.
     Note that ``bessel`` basis cannot be normalized.
@@ -37,11 +37,11 @@ def soft_one_hot_linspace(x, start, end, number, basis='gaussian', endpoint=True
     number : int
         number of basis functions :math:`N`
 
-    basis : {'gaussian', 'cosine', 'fourier', 'bessel', 'smooth_finite'}
+    basis : {'gaussian', 'cosine', 'smooth_finite', 'fourier', 'bessel'}
         choice of basis family; note that due to the :math:`1/x` term, ``bessel`` basis does not satisfy the normalization of other basis choices
 
-    endpoint : bool
-        if ``endpoint=False`` then for all :math:`x` outside of the interval defined by ``(start, end)``, :math:`\forall i, \; f_i(x) \approx 0`
+    cutoff : bool
+        if ``cutoff=True`` then for all :math:`x` outside of the interval defined by ``(start, end)``, :math:`\forall i, \; f_i(x) \approx 0`
 
     Returns
     -------
@@ -60,46 +60,43 @@ def soft_one_hot_linspace(x, start, end, number, basis='gaussian', endpoint=True
 
     .. jupyter-execute::
 
+        bases = ['gaussian', 'cosine', 'smooth_finite', 'fourier', 'bessel']
         x = torch.linspace(-1.0, 2.0, 100)
-        plt.plot(x, soft_one_hot_linspace(x, -0.5, 1.5, 3, 'gaussian', endpoint=True));
 
     .. jupyter-execute::
 
-        plt.plot(x, soft_one_hot_linspace(x, -0.5, 1.5, 3, 'smooth_finite', endpoint=False));
+        fig, axss = plt.subplots(len(bases), 2, figsize=(9, 6), sharex=True, sharey=True)
+
+        for axs, b in zip(axss, bases):
+            for ax, c in zip(axs, [True, False]):
+                plt.sca(ax)
+                plt.plot(x, soft_one_hot_linspace(x, -0.5, 1.5, number=4, basis=b, cutoff=c))
+                plt.plot([-0.5]*2, [-2, 2], 'k-.')
+                plt.plot([1.5]*2, [-2, 2], 'k-.')
+                plt.title(f"{b}" + (" with cutoff" if c else ""))
+
+        plt.ylim(-1, 1.5)
+        plt.tight_layout()
 
     .. jupyter-execute::
 
-        plt.plot(x, soft_one_hot_linspace(x, -0.5, 1.5, 3, 'cosine', endpoint=True));
+        fig, axss = plt.subplots(len(bases), 2, figsize=(9, 6), sharex=True, sharey=True)
 
-    .. jupyter-execute::
+        for axs, b in zip(axss, bases):
+            for ax, c in zip(axs, [True, False]):
+                plt.sca(ax)
+                plt.plot(x, soft_one_hot_linspace(x, -0.5, 1.5, number=4, basis=b, cutoff=c).pow(2).sum(1))
+                plt.plot([-0.5]*2, [-2, 2], 'k-.')
+                plt.plot([1.5]*2, [-2, 2], 'k-.')
+                plt.title(f"{b}" + (" with cutoff" if c else ""))
 
-        plt.plot(x, soft_one_hot_linspace(x, -0.5, 1.5, 3, 'cosine', endpoint=False));
-
-    .. jupyter-execute::
-
-        plt.plot(x, soft_one_hot_linspace(x, -0.5, 1.5, 3, 'fourier', endpoint=False));
-
-    .. jupyter-execute::
-
-        plt.plot(x, soft_one_hot_linspace(x, -0.5, 1.5, 3, 'fourier', endpoint=True));
-
-    .. jupyter-execute::
-
-        plt.plot(x, soft_one_hot_linspace(x, -0.5, 1.5, 3, 'bessel', endpoint=False));
-
-    .. jupyter-execute::
-
-        plt.plot(x, soft_one_hot_linspace(x, -0.5, 1.5, 3, 'bessel', endpoint=True));
-
-    .. jupyter-execute::
-
-        for basis in ['gaussian', 'cosine', 'fourier', 'smooth_finite']:
-            for endpoint in [False, True]:
-                y = soft_one_hot_linspace(x, -0.5, 1.5, 4, basis, endpoint)
-                plt.plot(x, y.pow(2).sum(1), label=f"{basis} {'endpoint' if endpoint else ''}")
-        plt.legend();
+        plt.ylim(0, 2)
+        plt.tight_layout()
     """
-    if endpoint:
+    if cutoff not in [True, False]:
+        raise ValueError("cutoff must be specified")
+
+    if not cutoff:
         values = torch.linspace(start, end, number, dtype=x.dtype, device=x.device)
         step = values[1] - values[0]
     else:
@@ -120,7 +117,7 @@ def soft_one_hot_linspace(x, start, end, number, basis='gaussian', endpoint=True
 
     if basis == 'fourier':
         x = (x[..., None] - start) / (end - start)
-        if endpoint:
+        if not cutoff:
             i = torch.arange(0, number, dtype=x.dtype, device=x.device)
             return torch.cos(math.pi * i * x) / math.sqrt(0.25 + number / 2)
         else:
@@ -133,7 +130,9 @@ def soft_one_hot_linspace(x, start, end, number, basis='gaussian', endpoint=True
         bessel_roots = torch.arange(1, number + 1, dtype=x.dtype, device=x.device) * math.pi
         out = math.sqrt(2 / c) * torch.sin(bessel_roots * x / c) / x
 
-        if endpoint:
+        if not cutoff:
             return out
         else:
             return out * ((x / c) < 1) * (0 < x)
+
+    raise ValueError(f"basis=\"{basis}\" is not a valid entry")