+++ e3nn/nn/models/v2103/gate_points_networks.py
@@ -75,8 +75,8 @@ class SimpleNetwork(torch.nn.Module):
             0.0,
             self.max_radius,
             self.number_of_basis,
-            basis='cosine',  # the cosine basis with endpoint = True goes to zero at max_radius
-            endpoint=False,  # no need for an additional smooth cutoff
+            basis='cosine',  # the cosine basis with cutoff = True goes to zero at max_radius
+            cutoff=True,  # no need for an additional smooth cutoff
         ).mul(self.number_of_basis**0.5)
 
         # Node attributes are not used here
@@ -152,8 +152,8 @@ class NetworkForAGraphWithAttributes(torch.nn.Module):
             0.0,
             self.max_radius,
             self.number_of_basis,
-            basis='cosine',  # the cosine basis with endpoint = True goes to zero at max_radius
-            endpoint=False,  # no need for an additional smooth cutoff
+            basis='cosine',  # the cosine basis with cutoff = True goes to zero at max_radius
+            cutoff=True,  # no need for an additional smooth cutoff
         ).mul(self.number_of_basis**0.5)
 
         node_outputs = self.mp(data['node_input'], data['node_attr'], edge_src, edge_dst, edge_attr, edge_length_embedding)